rpm: Set executable bits on installed libraries
Set the executable bits on the libraries installed in libdir.  This
change is important because it causes 'rpmbuild' to generate Provide
tag metadata for the libraries in the package, which is necessary now
that some binaries in other packages have generated Requires tags for
libraries packaged in the base package.  'rpmbuild' will not generate
the Provides tag if the libraries lack executable permission.

This change is part of 3f7d8ec219e1aa04b6c0417ecf5e730d40b4f149
on master (the rest of that change isn't applicable to the 1.6 branch)

Change-Id: Ie95d0d5e16745cfbf0d2a733a9421f94a89216ff
Reviewed-on: http://gerrit.openafs.org/5204
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: BuildBot <buildbot@rampaginggeek.com>