Use rra-c-util m4 from src/external
This commit switches us over from using manually maintained copies
of Russ's rra-c-util m4 macros, to using ones that are automatically
imported from his upstream git tree.

The macro versioning in the OpenAFS tree was slightly confused. This
(approximately) updates us from using version 3.3 of the macros to
version 3.8. The signifcant changes are:

  *) Use PATH_KRB5_CONFIG rather than KRB5_CONFIG to set the location
     of the krb5-config file
  *) Use --with-gssapi-{include, lib} if given, rather than krb5-config
  *) Use --with-krb5-{include, lib} if given, rather than krb5-config
  *) Define HAVE_KERBEROS if we find a Kerberos library

The first, in particular, is likely to cause some confusion.

Change-Id: Ie7a1c0db31b32fc7ac9ad7bded2c4ffce3ac013f
Reviewed-on: http://gerrit.openafs.org/5418
Reviewed-by: Derrick Brashear <shadow@dementix.org>
Tested-by: BuildBot <buildbot@rampaginggeek.com>